+++ app/src/main/AndroidManifest.xml
@@ -10,11 +10,11 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".ui.MovieListActivity">
+        <activity android:name=".ui.activities.MovieListActivity">
         </activity>
-        <activity android:name=".ui.MovieDetailActivity">
+        <activity android:name=".ui.activities.MovieDetailActivity">
         </activity>
-        <activity android:name=".ui.MainActivity">
+        <activity android:name=".ui.activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
