+++ app/src/main/java/com/example/movie_list/ui/components/ListAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.movie_list.ui
+package com.example.movie_list.ui.components
 
 import trikita.anvil.RenderableAdapter
